pyproject.toml
[tool.poetry]
name = "workflow-automate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "api", from = "src" },
    { include = "orchestration", from = "src" },
    { include = "relay", from = "src" },
    { include = "shared", from = "src" },
    { include = "worker", from = "src" },
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = "*"
gunicorn = "*"
celery = {extras = ["redis"], version = "*"}
psycopg2-binary = "*"
redis = "*"
pika = "*"
boto3 = "*"
structlog = "*"
httpx = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
testcontainers = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
